refactor: stringlike
