+++ packages/encodable/src/parsers/scale/createScaleFromScaleConfig.ts
@@ -13,7 +13,7 @@ import {
 import {
   ScaleType,
   Value,
-  CategoricalScaleInput,
+  StringLike,
   AllScale,
   ScaleConfig,
   LinearScaleConfig,
@@ -68,15 +68,15 @@ function createScaleFromScaleConfig<Output extends Value>(
 
 function createScaleFromScaleConfig<Output extends Value>(
   config: OrdinalScaleConfig<Output> | BinOrdinalScaleConfig<Output>,
-): ScaleOrdinal<CategoricalScaleInput, Output>;
+): ScaleOrdinal<StringLike, Output>;
 
 function createScaleFromScaleConfig<Output extends Value>(
   config: PointScaleConfig,
-): ScalePoint<CategoricalScaleInput>;
+): ScalePoint<StringLike>;
 
 function createScaleFromScaleConfig<Output extends Value>(
   config: BandScaleConfig,
-): ScaleBand<CategoricalScaleInput>;
+): ScaleBand<StringLike>;
 
 function createScaleFromScaleConfig<Output extends Value>(
   config: ScaleConfig<Output>,
@@ -92,7 +92,7 @@ function createScaleFromScaleConfig<Output extends Value>(config: ScaleConfig<Ou
     const scheme = 'scheme' in config ? config.scheme : undefined;
     const resolve = OptionsManager.getCategoricalColorScaleResolver();
 
-    let colorScale: ScaleOrdinal<CategoricalScaleInput, string>;
+    let colorScale: ScaleOrdinal<StringLike, string>;
     if (typeof scheme === 'undefined') {
       colorScale = resolve({});
     } else if (isSchemeParams(scheme)) {
@@ -101,7 +101,7 @@ function createScaleFromScaleConfig<Output extends Value>(config: ScaleConfig<Ou
       colorScale = resolve({ name: scheme });
     }
 
-    const castedColorScale = (colorScale as unknown) as ScaleOrdinal<CategoricalScaleInput, Output>;
+    const castedColorScale = (colorScale as unknown) as ScaleOrdinal<StringLike, Output>;
     applyDomain(config, castedColorScale);
 
     return castedColorScale;
