+++ packages/encodable/src/types/internal/Base.ts
@@ -4,9 +4,6 @@ export type Unarray<T> = T extends Array<infer U> ? U : T;
 /** T or an array of T */
 export type MayBeArray<T> = T | T[];
 
-/** A value that has .toString() function */
-export type StringLike = { toString(): string };
-
 /** Make some fields that might have been optional become required fields */
 export type RequiredSome<T, RequiredFields extends keyof T> = {
   [Field in Exclude<keyof T, RequiredFields>]?: T[Field];
